Add default error handler and custom exception
Frankt now provides a default error handler for channels that can be
overriden when desired. The default error handler simply logs the
received error and pushes a message to the channel.

There is also a custom ConfigurationException module that will allow
users to identify which errors come from Frankt configuration and
which ones come from custom handlers.